+++ test/entrypoint/srv_not_initialized/cartridge_init.lua
@@ -7,29 +7,16 @@ local log = require('log')
 local errors = require('errors')
 local cartridge = require('cartridge')
 
+if package.setsearchroot ~= nil then
+    package.setsearchroot()
+else
+    package.path = package.path .. debug.sourcedir() .. "/?.lua;"
+end
+
 package.preload['customers-storage'] = function()
     return {
         role_name = 'customers-storage',
-        init = function()
-            local customers_space = box.schema.space.create('customers', {
-                format = {
-                    {name = 'id', type = 'unsigned'},
-                    {name = 'bucket_id', type = 'unsigned'},
-                    {name = 'name', type = 'string'},
-                    {name = 'age', type = 'number'},
-                },
-                if_not_exists = true,
-            })
-            customers_space:create_index('id', {
-                parts = { {field = 'id' } },
-                if_not_exists = true,
-            })
-            customers_space:create_index('bucket_id', {
-                parts = { {field = 'bucket_id' } },
-                unique = false,
-                if_not_exists = true,
-            })
-        end,
+        init = require('storage_init'),
         dependencies = {'cartridge.roles.vshard-storage'},
     }
 end